cloud_function_crawl/requirements.txt
functions-framework==3.5.0
google-cloud-bigquery==3.17.2
beautifulsoup4==4.12.2
requests==2.31.0
